# docs/swagger.yaml
basePath: /
definitions:
...
  dto.CreateTaskRequest:
    properties:
      description:
        type: string
      group_id:
        type: integer
      title:
        type: string
    required:
    - group_id
    - title
    type: object
...
  dto.TaskDTO:
    properties:
      created_at:
        type: string
      description:
        type: string
      group_id:
        type: integer
      id:
        type: integer
      is_verified:
        type: boolean
      title:
        type: string
      user_id:
        type: integer
      username:
        type: string
    type: object
...
  dto.VerificationRequest:
    properties:
      is_verified:
        type: boolean
    required:
    - is_verified
    type: object
...
host: 185.221.155.133:8080
info:
  contact:
    email: vasyutin.i.a@edu.mirea.ru, saakovi.d.@edu.mirea.ru, pronin.p.v@edu.mirea.ru
    name: Иван Васютин, Павел Пронин, Давит Саакови
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: СВАГА
  version: "1.0"
paths:
...
  /api/tasks:
    get:
      consumes:
      - application/json
      description: Get a list of tasks for the specified group
      parameters:
      - description: Group ID
        in: query
        name: group_id
        required: true
        type: integer
      - description: Bearer JWT
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.TaskDTO'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get tasks in a group
      tags:
      - tasks
    post:
      consumes:
      - application/json
      description: Create a new task in the specified group
      parameters:
      - description: Task data
        in: body
        name: task
        required: true
        schema:
          $ref: '#/definitions/dto.CreateTaskRequest'
      - description: Bearer JWT
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a new task
      tags:
      - tasks
  /api/tasks/{id}:
    get:
      consumes:
      - application/json
      description: Get details of a task by ID, if user is a member of the task's
        group
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: integer
      - description: Bearer JWT
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.TaskDTO'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get a specific task
      tags:
      - tasks
  /api/tasks/{id}/verify:
    patch:
      consumes:
      - application/json
      description: Verify or deny a task's legitimacy
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: integer
      - description: Verification Status
        in: body
        name: verify
        required: true
        schema:
          $ref: '#/definitions/dto.VerificationRequest'
      - description: Bearer JWT
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Verify a task
      tags:
      - tasks
  /api/tasks/my-groups:
    get:
      consumes:
      - application/json
      description: Get a list of tasks from all groups the user is a member of
      parameters:
      - description: Bearer JWT
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.TaskDTO'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get tasks from all user's groups
      tags:
      - tasks
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"